pyproject.toml
[project]
name = "django-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "drf-spectacular[sidecar]==0.28.0",
    "markdown>=3.7",
]

[dependency-groups]
dev = [
    "pytest-django>=4.9.0",
    "ruff>=0.8.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"
python_files = [
    "tests/test_*.py",
    "tests.py",
]
addopts = "-s --reuse-db"
django_debug_mode = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
